Security

+++ src/main/java/com/skypedal/skypedal_backend/dto/UsersRewardsDTO.java
@@ -11,13 +11,13 @@ public class UsersRewardsDTO {
     private boolean hasUsed;
 
     private Integer rewardId;
-    private Integer userId;
+    private Long userId;
 
     public UsersRewardsDTO() {
         super();
     }
 
-    public UsersRewardsDTO(Integer id, String dateRedeemed, String dateExpiry, boolean hasUsed, Integer rewardId, Integer userId) {
+    public UsersRewardsDTO(Integer id, String dateRedeemed, String dateExpiry, boolean hasUsed, Integer rewardId, Long userId) {
         this.id = id;
         this.dateRedeemed = dateRedeemed;
         this.dateExpiry = dateExpiry;
@@ -76,11 +76,11 @@ public class UsersRewardsDTO {
         this.rewardId = rewardId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 }
